Revert "Rename matchers packages to singular case (#736)"
This reverts commit e27ac9596ae4384cab57b54338a93bf4850a64d2.
