+++ kotlintest-assertions/src/commonMain/kotlin/io/kotlintest/matchers/floats/matchers.kt
@@ -1,4 +1,4 @@
-package io.kotlintest.matchers.float
+package io.kotlintest.matchers.floats
 
 import io.kotlintest.Matcher
 import io.kotlintest.Result
