+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/ContinuallyTest.kt
@@ -1,7 +1,7 @@
 package com.sksamuel.kotlintest
 
 import io.kotlintest.continually
-import io.kotlintest.matchers.integer.shouldBeLessThan
+import io.kotlintest.matchers.integers.shouldBeLessThan
 import io.kotlintest.matchers.string.shouldMatch
 import io.kotlintest.shouldBe
 import io.kotlintest.shouldThrowExactly
