+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/DslExamples.kt
@@ -1,7 +1,7 @@
 package com.sksamuel.kotlintest
 
-import io.kotlintest.matchers.integer.beGreaterThan
-import io.kotlintest.matchers.integer.beLessThan
+import io.kotlintest.matchers.integers.beGreaterThan
+import io.kotlintest.matchers.integers.beLessThan
 import io.kotlintest.matchers.singleElement
 import io.kotlintest.should
 import io.kotlintest.shouldBe
