+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/matchers/MatcherDsl.kt
@@ -1,7 +1,7 @@
 package com.sksamuel.kotlintest.matchers
 
-import io.kotlintest.matchers.double.exactly
-import io.kotlintest.matchers.double.plusOrMinus
+import io.kotlintest.matchers.doubles.exactly
+import io.kotlintest.matchers.doubles.plusOrMinus
 import io.kotlintest.matchers.file.haveExtension
 import io.kotlintest.matchers.string.endWith
 import io.kotlintest.matchers.string.startWith
@@ -13,10 +13,10 @@ import java.io.FileNotFoundException
 
 fun matcherDsl() {
 
-  // example of a tolerance match for double
+  // example of a tolerance match for doubles
   1.0 shouldBe (1.3 plusOrMinus 0.2)
 
-  // example of an exact match for double
+  // example of an exact match for doubles
   1.0 shouldBe exactly(1.0)
 
   // example of expecting an exception
