+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/matchers/SoftAssertionsTest.kt
@@ -3,13 +3,13 @@ package com.sksamuel.kotlintest.matchers
 import io.kotlintest.assertSoftly
 import io.kotlintest.matchers.collections.containExactly
 import io.kotlintest.matchers.collections.shouldNotContainExactly
-import io.kotlintest.matchers.double.negative
-import io.kotlintest.matchers.double.positive
-import io.kotlintest.matchers.double.shouldBeNegative
+import io.kotlintest.matchers.doubles.negative
+import io.kotlintest.matchers.doubles.positive
+import io.kotlintest.matchers.doubles.shouldBeNegative
 import io.kotlintest.matchers.haveKey
-import io.kotlintest.matchers.integer.beLessThan
-import io.kotlintest.matchers.integer.shouldBeLessThan
-import io.kotlintest.matchers.integer.shouldBePositive
+import io.kotlintest.matchers.integers.beLessThan
+import io.kotlintest.matchers.integers.shouldBeLessThan
+import io.kotlintest.matchers.integers.shouldBePositive
 import io.kotlintest.matchers.string.contain
 import io.kotlintest.matchers.string.endWith
 import io.kotlintest.matchers.string.shouldNotEndWith
