+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/matchers/doubles/DoubleMatchersTest.kt
@@ -1,33 +1,33 @@
 package com.sksamuel.kotlintest.matchers.doubles
 
-import io.kotlintest.matchers.double.beGreaterThan
-import io.kotlintest.matchers.double.beGreaterThanOrEqualTo
-import io.kotlintest.matchers.double.beLessThan
-import io.kotlintest.matchers.double.beLessThanOrEqualTo
-import io.kotlintest.matchers.double.between
-import io.kotlintest.matchers.double.exactly
-import io.kotlintest.matchers.double.gt
-import io.kotlintest.matchers.double.gte
-import io.kotlintest.matchers.double.lt
-import io.kotlintest.matchers.double.lte
-import io.kotlintest.matchers.double.negative
-import io.kotlintest.matchers.double.positive
-import io.kotlintest.matchers.double.shouldBeBetween
-import io.kotlintest.matchers.double.shouldBeExactly
-import io.kotlintest.matchers.double.shouldBeGreaterThan
-import io.kotlintest.matchers.double.shouldBeGreaterThanOrEqual
-import io.kotlintest.matchers.double.shouldBeLessThan
-import io.kotlintest.matchers.double.shouldBeLessThanOrEqual
-import io.kotlintest.matchers.double.shouldBeNegative
-import io.kotlintest.matchers.double.shouldBePositive
-import io.kotlintest.matchers.double.shouldNotBeBetween
-import io.kotlintest.matchers.double.shouldNotBeExactly
-import io.kotlintest.matchers.double.shouldNotBeGreaterThan
-import io.kotlintest.matchers.double.shouldNotBeGreaterThanOrEqual
-import io.kotlintest.matchers.double.shouldNotBeLessThan
-import io.kotlintest.matchers.double.shouldNotBeLessThanOrEqual
-import io.kotlintest.matchers.double.shouldNotBeNegative
-import io.kotlintest.matchers.double.shouldNotBePositive
+import io.kotlintest.matchers.doubles.beGreaterThan
+import io.kotlintest.matchers.doubles.beGreaterThanOrEqualTo
+import io.kotlintest.matchers.doubles.beLessThan
+import io.kotlintest.matchers.doubles.beLessThanOrEqualTo
+import io.kotlintest.matchers.doubles.between
+import io.kotlintest.matchers.doubles.exactly
+import io.kotlintest.matchers.doubles.gt
+import io.kotlintest.matchers.doubles.gte
+import io.kotlintest.matchers.doubles.lt
+import io.kotlintest.matchers.doubles.lte
+import io.kotlintest.matchers.doubles.negative
+import io.kotlintest.matchers.doubles.positive
+import io.kotlintest.matchers.doubles.shouldBeBetween
+import io.kotlintest.matchers.doubles.shouldBeExactly
+import io.kotlintest.matchers.doubles.shouldBeGreaterThan
+import io.kotlintest.matchers.doubles.shouldBeGreaterThanOrEqual
+import io.kotlintest.matchers.doubles.shouldBeLessThan
+import io.kotlintest.matchers.doubles.shouldBeLessThanOrEqual
+import io.kotlintest.matchers.doubles.shouldBeNegative
+import io.kotlintest.matchers.doubles.shouldBePositive
+import io.kotlintest.matchers.doubles.shouldNotBeBetween
+import io.kotlintest.matchers.doubles.shouldNotBeExactly
+import io.kotlintest.matchers.doubles.shouldNotBeGreaterThan
+import io.kotlintest.matchers.doubles.shouldNotBeGreaterThanOrEqual
+import io.kotlintest.matchers.doubles.shouldNotBeLessThan
+import io.kotlintest.matchers.doubles.shouldNotBeLessThanOrEqual
+import io.kotlintest.matchers.doubles.shouldNotBeNegative
+import io.kotlintest.matchers.doubles.shouldNotBePositive
 import io.kotlintest.properties.Gen
 import io.kotlintest.properties.assertAll
 import io.kotlintest.should
