+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/matchers/numerics/IntMatchersTest.kt
@@ -1,22 +1,22 @@
 package com.sksamuel.kotlintest.matchers.numerics
 
-import io.kotlintest.matchers.integer.beEven
-import io.kotlintest.matchers.integer.beGreaterThan
-import io.kotlintest.matchers.integer.beGreaterThanOrEqualTo
-import io.kotlintest.matchers.integer.beInRange
-import io.kotlintest.matchers.integer.beLessThan
-import io.kotlintest.matchers.integer.beLessThanOrEqualTo
-import io.kotlintest.matchers.integer.beOdd
-import io.kotlintest.matchers.integer.between
-import io.kotlintest.matchers.integer.lt
-import io.kotlintest.matchers.integer.lte
-import io.kotlintest.matchers.integer.shouldBeBetween
-import io.kotlintest.matchers.integer.shouldBeGreaterThan
-import io.kotlintest.matchers.integer.shouldBeGreaterThanOrEqual
-import io.kotlintest.matchers.integer.shouldBeLessThan
-import io.kotlintest.matchers.integer.shouldBeLessThanOrEqual
-import io.kotlintest.matchers.integer.shouldBeNegative
-import io.kotlintest.matchers.integer.shouldBePositive
+import io.kotlintest.matchers.integers.beEven
+import io.kotlintest.matchers.integers.beGreaterThan
+import io.kotlintest.matchers.integers.beGreaterThanOrEqualTo
+import io.kotlintest.matchers.integers.beInRange
+import io.kotlintest.matchers.integers.beLessThan
+import io.kotlintest.matchers.integers.beLessThanOrEqualTo
+import io.kotlintest.matchers.integers.beOdd
+import io.kotlintest.matchers.integers.between
+import io.kotlintest.matchers.integers.lt
+import io.kotlintest.matchers.integers.lte
+import io.kotlintest.matchers.integers.shouldBeBetween
+import io.kotlintest.matchers.integers.shouldBeGreaterThan
+import io.kotlintest.matchers.integers.shouldBeGreaterThanOrEqual
+import io.kotlintest.matchers.integers.shouldBeLessThan
+import io.kotlintest.matchers.integers.shouldBeLessThanOrEqual
+import io.kotlintest.matchers.integers.shouldBeNegative
+import io.kotlintest.matchers.integers.shouldBePositive
 import io.kotlintest.should
 import io.kotlintest.shouldBe
 import io.kotlintest.shouldNot
