+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/matchers/numerics/LongMatchersTest.kt
@@ -4,9 +4,9 @@ import io.kotlintest.matchers.beGreaterThan
 import io.kotlintest.matchers.beGreaterThanOrEqualTo
 import io.kotlintest.matchers.beLessThan
 import io.kotlintest.matchers.beLessThanOrEqualTo
-import io.kotlintest.matchers.long.between
-import io.kotlintest.matchers.long.shouldBeNegative
-import io.kotlintest.matchers.long.shouldBePositive
+import io.kotlintest.matchers.longs.between
+import io.kotlintest.matchers.longs.shouldBeNegative
+import io.kotlintest.matchers.longs.shouldBePositive
 import io.kotlintest.should
 import io.kotlintest.shouldBe
 import io.kotlintest.shouldThrow
