+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/properties/GenTest.kt
@@ -8,10 +8,10 @@ import io.kotlintest.matchers.boolean.shouldBeTrue
 import io.kotlintest.matchers.collections.contain
 import io.kotlintest.matchers.collections.shouldContain
 import io.kotlintest.matchers.collections.shouldContainAll
-import io.kotlintest.matchers.double.shouldBeGreaterThanOrEqual
-import io.kotlintest.matchers.double.shouldBeLessThanOrEqual
-import io.kotlintest.matchers.float.shouldBeGreaterThanOrEqual
-import io.kotlintest.matchers.float.shouldBeLessThanOrEqual
+import io.kotlintest.matchers.doubles.shouldBeGreaterThanOrEqual
+import io.kotlintest.matchers.doubles.shouldBeLessThanOrEqual
+import io.kotlintest.matchers.floats.shouldBeGreaterThanOrEqual
+import io.kotlintest.matchers.floats.shouldBeLessThanOrEqual
 import io.kotlintest.matchers.gte
 import io.kotlintest.matchers.lt
 import io.kotlintest.matchers.string.include
@@ -26,6 +26,7 @@ import io.kotlintest.specs.WordSpec
 import io.kotlintest.tables.headers
 import io.kotlintest.tables.row
 import io.kotlintest.tables.table
+import org.junit.jupiter.api.assertAll
 import java.time.Duration
 import java.time.LocalDate
 import java.time.LocalDateTime
