+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/properties/shrinking/StringShrinkerTest.kt
@@ -1,6 +1,6 @@
 package com.sksamuel.kotlintest.properties.shrinking
 
-import io.kotlintest.matchers.integer.shouldBeLessThan
+import io.kotlintest.matchers.integers.shouldBeLessThan
 import io.kotlintest.matchers.string.shouldEndWith
 import io.kotlintest.matchers.string.shouldHaveLength
 import io.kotlintest.matchers.string.shouldNotContain
