+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/specs/feature/FeatureSpecTest.kt
@@ -1,6 +1,6 @@
 package com.sksamuel.kotlintest.specs.feature
 
-import io.kotlintest.matchers.integer.shouldBeLessThan
+import io.kotlintest.matchers.integers.shouldBeLessThan
 import io.kotlintest.specs.FeatureSpec
 
 class FeatureSpecTest : FeatureSpec() {
