+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/specs/wordspec/WordSpecTest.kt
@@ -1,6 +1,6 @@
 package com.sksamuel.kotlintest.specs.wordspec
 
-import io.kotlintest.matchers.integer.shouldBeGreaterThan
+import io.kotlintest.matchers.integers.shouldBeGreaterThan
 import io.kotlintest.specs.WordSpec
 
 class WordSpecTest : WordSpec() {
